[class] build the class brand and private AccessorPairs eagerly
Once an accessor is declared, later access to it in computed
property keys should throw a TypeError (for brand checks)
instead of a ReferenceError. This patch implements it by building
the AccessorPair eagerly once *any* of the component is visited
so that the hole check can be passed, and building
the private class brand before visiting class member keys so that
the brand can be checked in computed property keys.

This means that we cannot create the complementary getter and setter
together, and have to split the creation of the AccessorPair and the
assignment of at least one of its component, in case one component
is accessed before the other one is declared:

class C {
  get #c() {  }
  [this.#c] = 1;  // TypeError
  set #c(val) {  }
}

Note that this is still a ReferenceError:

class C {
  [this.#c] = 1;  // ReferenceError
  get #c() {  }
  set #c(val) {  }
}

Refs: https://github.com/tc39/proposal-class-fields/issues/263

Bug: v8:8330