[runtime] Add DefineOwn IC/opcode/etc for class fields
Adds some new opcodes/IC kinds for storing defining class properties.

StaKeyedPropertyAsDefine is used for private fields.
It provides an extra bit of information (that this is a define
operation, with slightly different throwing behaviour than a store operation, in
the slow case!). The difference is very minor, however this was chosen
in lieu of adding additional parameters to keyed store ops in general.

StoreOwnIC is sort of a replacement for %CreateDataProperty() --- This
new method was created in part because Store*InLiteral seems to depend
on the map already being populated by an object boilerplate.

Ultimately, this change does produce a pretty significant improvement in
scores on the new microbenchmark, which should indicate some improvement
in applications which use the mixin pattern with classes. It shouldn't
produce any significant change for other accesses to class fields.

TurboFan support has been added, perhaps not in the best way possible. It
does at the very least ensure that on the generic path, the correct IC stub
call is emitted. I haven't yet found evidence that native context optimization
does anything useful with it, though.

R=leszeks@chromium.org, adamk@chromium.org, rmcilroy@chromium.org

Bug:v8:9888
Change-Id: Idcc947585c0e612f9e8533aa4e2e0f8f0df8875d