[class] implement reparsing of class instance member initializers
Previously, since the source code for the synthetic class instance
member initializer function was recorded as the entire class body,
there was no way to reparse the class and recompile the initializer
function. It was working for most use cases because the code for the
initializer function was generated eagarly and it was usually alive as
long as the class was alive, so the initializer wouldn't normally be
lazily parsed. This didn't work, however, when the class was
snapshotted with v8::SnapshotCreator::FunctionCodeHandling::kClear,
becuase then we needed to recompile the initializer when the class
was instantiated. This patch implements the reparsing so that
these classes can work with FunctionCodeHandling::kClear.

This patch refactors ParserBase::ParseClassLiteral() so that we can
reuse it for both parsing the class body normally and reparsing it
to collect initializers. When reparsing the synthetic initializer
function, we rewind the scanner to the beginning of the class, and
parse the class body to collect the initializers. We use the pre-parser
to parse any member that's not a instance field initializer, and revert
to the full parser for initializers. When declaring a class member
or the home objects, instead of declaring the variables, we simply
look them up from the scope info and put them back into the variable
maps.

A few notable changes:

- Added AstRawString::GetInternalized() to create internalized copies
  of variable names in AstRawStrings, this is necessary to look them up
  from the scope info.
- Moved the parsing mode enum from the full parser to ParserBase
  for code reuse in ParserBase::ParseClassLiteral()
- Extended the source range of the initializer function to cover the
  entire class so that we can rewind the scanner to parse the class
  body to collect initializers (previously, it starts from the first
  field initializer and ends at the last initializer). This resulted
  some expectation changes in the debugger tests, though the
  initializers remain debuggable.

Bug: v8:10704
Change-Id: Ifb6431a1447d8844f2a548283d59158742fe9027