[snapshot] handle configured Error.stackTraceLimit in snapshot
Previously, deserializing a snapshot where the Error.stackTraceLimit
is configured results in a DCHECK because the bootstrapper uses
JSObject::AddProperty() to install it at startup unconditionally.

This fixes that by skipping the installation when we find that the
Error.stackTraceLimit already exists, which is similar to how
we handle named properties on the global object in a snapshot.

Change-Id: I4f399f85dc07a27534ab99fbeb7428af24aee621