[class] handle reparsing of initializers during error rendering
The outer scope of the initializer function scope is a script
scope during reparsing for error rendering, in that case
we cannot reparse them properly, so just parse them as normal
functions, then it would fail and the error rendering would
not attempt to analyze the AST for destructring errors.

https://chromium-review.googlesource.com/c/v8/v8/+/2988830
implemented reparsing in case the outer scope **is** a class
scope, and this patch just fixes the regression when the outer
scope isn't a class scope by restoring the old behavior, which
is just an error message not as specific but still informative.