[api] add Module::GetResourceName()
This allows embedders to retrieve the resource name of a module
directly instead of having to keep track of them on the side.
If the module is a Source Text Module, returns the name
that was passed by the embedder as resource_name to the ScriptOrigin.
If it's a Synthetic Module, returns the module_name passed to
CreateSyntheticModule().